OLMIS-1079, remove logging configuration file - not working in CI
It appears that docker-compose host mounting this file into remote hosts doesn't work.  This would make sense as a host mount means the docker host.  When that host is a remote ec2 instance, as it is with our CI/CD setup, then the file we thought we were host mounting is actually on the CI/CD box, and not the remote docker host.  We SHOULD AVOID setting up any host mounts in the future.