OLMIS-1738, change keep/wipe to be db centric
Whereas before this was an option that effected docker volumes (which also had the side-effect of effecting the included database container), this now is database centric by utilizing the setting a service has to keep or wipe its db on startup.  This should be the desired approach, as clearing log volumes (the only other use for docker volumes here) doesn't seem very desirable here.