Add NiFi Registry Deployment (#7)
* OLMIS-5086: Add NiFi Registry Deployment

Since OpenLMIS needs a shared NiFi Registry instance, add deployment
scripts for it (to be used with Jenkins).

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5086: Add Terraform Def For NiFi Reg Deploy

Since the shared NiFi Registry instance will be deployed on an AWS
instance, add the definitions for that setup in Terraform files.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5086: Add Docker Install Auto Scripts

Since the NiFi Registry host will have to have the Docker Daemon
running, create Ansible scripts for installing Docker and Docker Compose
on that host.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5231: Use Ansible Tasks For TLS Gen Process

Since Ansible is unable to generate enough entropy when running the
shell script for generating the TLS files in the Docker role, switch to
generating the TLS files using Ansible tasks.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5231: Add Ansible pip Dependencies

Since a bunch of people will be using the defined Ansible playbooks, add
a file with the required pip dependencies. Also add documentation on how
to install the pip and Ansible Galaxy dependencies.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* Fix formatting for NiFi Registry Terraform

Do this by running:
  terraform fmt provision/terraform/modules/nifi-registry
  terraform fmt provision/terraform/reporting/nifi-registry

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5231: Setup Docker from within uat Terraform

Run the Ansible playbook for setting up Docker and Docker Compose from
within the uat Terraform definitions.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* Fix formatting for uat Terraform

Do this by running:
  terraform fmt provision/terraform/modules/uat
  terraform fmt provision/terraform/uat

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* Rename Terraform Var docker-https-port

Rename the Terraform variable docker-https-port to something more
appropriate (docker-tls-port).

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5231: Add DeployGroup EC2 Tag to uat Build

Since the Docker dynamic inventory will only run on hosts that have the
EC2 tag DeployGroup with a value of docker-hosts, add it to the uat
deployments.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5289, add benin-demo env

* OLMIS-5289, start README for how to use Terraform and Ansible for new env

* OLMIS-5086: Use openlmis-reporting SSH Key

Use the openlmis-reporting SSH key (whose private key is backed up in
the aws-instance-keys S3 bucket) in the NiFi Registry deployment.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5231: Standardise Terraform AWS Cred Vars

Use the same variable names for the Terraform variables to hold the AWS
key ID and AWS secret key across the UAT and NiFi Registry setups. This
allows for users to get the values from the same Terraform environment
variable.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-5231: Add Terraform Setup Instructions

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>