OLMIS-6425 Add security group id variable to override
Not all instances may be on the same security group. This allows Casper instance to be on a different security group, with more open inbound ports. More ports were opened for all instances because it is unclear how to add more listener ports to each individual instance. Seems like listeners can be set with a list (https://github.com/hashicorp/terraform/issues/9807), which could be a variable, but a default list of listeners cannot have interpolations (https://github.com/hashicorp/terraform/issues/18069), which may have been fixed in v0.12. For now, we just add the listeners.