OLMIS-6496 Open NiFi registry ports 18080, 18443
These ports are used already by the nifi-registry.openlmis.org server,
but the ELB did not need listeners for them because the ELB is actually
being bypassed for that server: the DNS entry for
nifi-registry.openlmis.org points directly to the instance, not the load
balancer. Since the Casper NiFi registry has its DNS record pointing to
its load balancer, we need these listeners.